extrac common functionality between build/merge
